Add `pull_build_push.sh` script to act as continuous deployment
This script is to be executed in a server that is always running in which executes the commands every minute. It will fetch any changes or commits pushed only specific to branch `origin/feat/multi-instance-gpu-ddp`. This will be only be updated for every modification related to the script, Dockerfile, and python scripts in this directory.